Refactor the section grouping function
